requirements.txt
appdirs==1.4.3
dj-database-url==0.4.1
Django==1.11.6
gunicorn==19.6.0
httplib2==0.10.3
oauth2client==4.0.0
packaging==16.8
psycopg2==2.6.2
pyasn1==0.2.3
pyasn1-modules==0.0.8
pyparsing==2.2.0
python-wordpress-xmlrpc==2.3
pytz==2017.3
PyYAML==3.12
requests==2.9.1
rsa==3.4.2
six==1.10.0
whitenoise==3.2
xlrd==1.1.0
